Fits CES boundary models

The boundary models are given in Table 2 of Heun, et al "An Empirical Investigation of the Role of Energy in Economic Growth"

cesBoundaryModels(formula, data, nest, method = "nlm", subset = TRUE)

Arguments

formula
a CES formula in the form y ~ a + b + c + d + time

data
historical time series data

nest
a permutation of the integers 1 to the number of factors describing how to nest and permute the variables in the formula.

Value

a list of boundary models.

Examples

if (require(EconData) & require(dplyr)) {
  cesBoundaryModels(iGDP ~ iK + iL + iQp + iYear, data=Calvin %>% filter(Country=="US"), nest=c(1,2,3))
  cesBoundaryModels(iGDP ~ iK + iL + iQp + iYear, data=Calvin %>% filter(Country=="US"), nest=c(2,3,1))
  cesBoundaryModels(iGDP ~ iK + iL + iYear, data=Calvin %>% filter(Country=="US"), nest=c(1,2))
  }

#> Loading required package: EconData

#> Loading required package: dplyr

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:MacroGrowth’:
#> 
#>     count, do, tally

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

#> Warning: NaNs produced

#> $`01:1-1-`
#> 
#> Call:
#> plm(formula = log(iGDP) - log(iK) ~ iYear, data = sdata, params = c(delta_1 = 1, 
#> delta = 1), method = "nlm")
#> 
#> Coefficients:
#> logscale    lambda   delta_1     delta  
#> 0.009248  0.003537  1.000000  1.000000  
#> 
#> 
#> $`02:0-1-`
#> 
#> Call:
#> plm(formula = log(iGDP) - log(iL) ~ iYear, data = sdata, params = c(delta_1 = 0, 
#> delta = 1), method = "nlm")
#> 
#> Coefficients:
#> logscale    lambda   delta_1     delta  
#> -0.02344   0.01793   0.00000   1.00000  
#> 
#> 
#> $`04:-01-`
#> 
#> Call:
#> plm(formula = log(iGDP) - log(pmin(iK, iL)) ~ iYear, data = sdata, 
#>     params = c(sigma_1 = 0, delta = 1), method = "nlm")
#> 
#> Coefficients:
#> logscale    lambda   sigma_1     delta  
#> -0.02344   0.01793   0.00000   1.00000  
#> 
#> 
#> $`07:*N1-`
#> 
#> Call:
#> plm(formula = log(iGDP) - log(delta_1 * iK + (1 - delta_1) * 
#>     iL) ~ iYear, data = sdata, params = c(delta_1 = 0.5, delta = 1, 
#> rho_1 = -1), method = "nlm")
#> 
#> Coefficients:
#> logscale    lambda   delta_1     delta     rho_1  
#> -0.01612   0.01548   0.14223   1.00000  -1.00000  
#> 
#>