Fitting single factor models

Fitting single factor models

sfModel(formula, data, response, factor, time, constrained = FALSE,
  save.data = TRUE, correlation = NULL)

Arguments

formula
a formula of the form y ~ factor + time

data
a data frame in which formula is evaluated

constrained
a logical indicating whether the model parameters are constrained

correlation
an optional corStruct object describing the within-group correlation structure. See the documentation of corClasses for a description of the available corStruct classes. If a grouping variable is to be used, it must be specified in the form argument to the corStruct constructor. Defaults to NULL, corresponding to uncorrelated errors.

Value

an lm object with some additional attributes

Examples

US <- subset(Calvin, Country=="US")
sfModel(response = iGDP, factor=iK, time = iYear, data=US)

#> Error in log(NULL): non-numeric argument to mathematical function

sfModel(iGDP ~ iK + iYear, data=US)

#> Generalized least squares fit by REML
#>   Model: log(iGDP) ~ iK + iYear 
#>   Data: data 
#>   Log-restricted-likelihood: 49.12319
#> 
#> Coefficients:
#> (Intercept)          iK       iYear 
#>  0.03451906 -0.02663869  0.03028059 
#> 
#> Degrees of freedom: 32 total; 29 residual
#> Residual standard error: 0.03858439 

sfModel(response="iGDP", factor="iK", time="iYear", data=US)

#> Generalized least squares fit by REML
#>   Model: log(iGDP) ~ iK + iYear 
#>   Data: data 
#>   Log-restricted-likelihood: 49.12319
#> 
#> Coefficients:
#> (Intercept)          iK       iYear 
#>  0.03451906 -0.02663869  0.03028059 
#> 
#> Degrees of freedom: 32 total; 29 residual
#> Residual standard error: 0.03858439