Tells whether a CES model is within constraints

This function evaluates a fitted CES model for economic meaningfullness as expressed in constraints on delta_1, sigma_1 and, if present, delta_2, sigma_2, delta, and sigma. Specifically, this function checks whether 0 <= delta <= 1 and sigma >= 0

withinConstraints(model)

Arguments

model
the CES model to be evaluated

Value

TRUE if the fitted model is within the constraints, FALSE if not.

Note

Missing, NULL, and NA parameters are interpreted as meeting constraints.

If none of delta_1, sigma_1, delta_2, sigma_2, delta, and sigma is present, FALSE is returned.

This function relies upon the presence of the naturalCoeffs attribute to model. If naturalCoeffs is NULL, FALSE is returned.